+++ tests/test_point.py
@@ -13,51 +13,45 @@ from .conftest import isclose
 
 class TestPoint:
     @given_inferred
-    def test_coordinates(self, x: np.float64, y: np.float64) -> None:
+    def test_coordinates(self, x: float, y: float) -> None:
         p = Point(x, y)
         assert p.x == x
         assert p.y == y
 
     @given_inferred
-    def test_equality(self, x: np.float64, y: np.float64) -> None:
+    def test_equality(self, x: float, y: float) -> None:
         assert Point(x, y) == Point(x, y)
 
     @given_inferred
-    def test_adding(
-        self, x1: np.float64, x2: np.float64, y1: np.float64, y2: np.float64
-    ):
+    def test_adding(self, x1: float, x2: float, y1: float, y2: float):
         a = Point(x1, y1)
         b = Point(x2, y2)
         assert a + b == Point(x1 + x2, y1 + y2)
 
     @given_inferred
-    def test_translation(
-        self, x1: np.float64, x2: np.float64, y1: np.float64, y2: np.float64
-    ):
+    def test_translation(self, x1: float, x2: float, y1: float, y2: float):
         a = Point(x1, y1)
         b = Point(x2, y2)
         assert a + b == Point(x1 + x2, y1 + y2)
 
     @given_inferred
-    def test_subtraction(
-        self, x1: np.float64, x2: np.float64, y1: np.float64, y2: np.float64
-    ):
+    def test_subtraction(self, x1: float, x2: float, y1: float, y2: float):
         a = Point(x1, y1)
         b = Point(x2, y2)
         assert a - b == Point(x1 - x2, y1 - y2)
 
     @given_inferred
-    def test_multiplication(self, x: np.float64, y: np.float64, s: np.float64):
+    def test_multiplication(self, x: float, y: float, s: float):
         a = Point(x, y)
         assert a * s == Point(x * s, y * s)
 
     @given_inferred
-    def test_scale(self, x: np.float64, y: np.float64, s: np.float64):
+    def test_scale(self, x: float, y: float, s: float):
         a = Point(x, y)
         assert a.scale(s) == Point(x * s, y * s)
 
     @given_inferred
-    def test_abs(self, x: np.float64, y: np.float64):
+    def test_abs(self, x: float, y: float):
         assume(x * x != inf)
         assume(y * y != inf)
         a = Point(x, y)
@@ -79,7 +73,7 @@ class TestPoint:
         assert -np.pi <= angle <= np.pi
 
     @given_inferred
-    def test_unit_vector(self, x: np.float64, y: np.float64):
+    def test_unit_vector(self, x: float, y: float):
         a = Point(x, y)
         if isclose(abs(a), 0.0):
             with pytest.raises(ZeroDivisionError):
